requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
decorator==4.4.2
dj-database-url==0.5.0
Django==3.2.5
django-crispy-forms==1.12.0
django-heroku==0.0.0
gunicorn==20.1.0
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonpickle==2.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
networkx==2.5.1
numpy==1.21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
pypi-install==0.0.5
pytz==2021.1
pyvis==0.1.9
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
whitenoise==5.3.0
